Initial version of the implementation of Connect Four using the minimax algorithm
It works, but a lot of things are still open: a heuristic function to
evaluate the leaves of the search tree is required; multiple cores are
used and also benefit the analysis time, but the current implementation
is not really compatible with alpha-beta pruning (which is currently
missing, anyway). Besides these issues a large number of minor issues
is also missing.